Per thread logging (#168)
Problem

Rust code should be able to log to ccommon's logging facility to reduce the complexity of configuring embedded features.

Solution

This PR contains code that allows rust to create per-thread logger instances for use with the standard log crate. This gives us lock-free logging on the rust side, and allows the use of any library without needing to be concerned that that library or its dependencies will log from other threads besides the calling thread. This is a fair tradeoff between speed, safety, and ease of use.

Result

Users will configure the rust logging extension cc_log_rs with a directory and a filename prefix. The extension will create new, unique log files under that directory based on that filename prefix per thread on first use.