Improve the Rust build story (#240)
Added a top-level cargo manifest so it can be used as the build
tool for Rust bindings and wrappers.

Added updated `ccommon-sys` binding which uses cmake to drive its
build process. This simplifies the build file.

Changes the `.gitignore` to include the cargo build folder, lock
file, and the generated bindings.