restore cmake for rust libraries (#241)
* restore cmake for rust libraries

Previous PR removed cmake for the rust libraries. This creates a
problem while integrating the changes into the Pelikan repo.

Restored the files to their previous versions with minor changes
to reflect the new directory structure.

Now we can build the rust libraries with cmake or with cargo. This
should allow for integration into Pelikan repo.

* fix find for linux

* add cargo based osx build