Delete flaky assert within check_tcp (#253)
This assert has been causing issues for CI in pelikan. It also doesn't test
what we want it to test since it also measures the latency of starting a new
thread in addition to the time that thread spends sleeping. Beyond that, this
test has had flakiness issues before and the response each time has been to
just bump up the tolerance. That suggests that it's not worth having thise
assert since everyone will just ignore test failures anyway.

As such, this change removes the assert which should hopefully be enough to
avoid downstream test flakiness.